refactor: move GET API to [date] route
