+++ src/app/api/records/[date]/route.ts
@@ -10,12 +10,74 @@ import {
   todoSnapshotSchema,
 } from "@/schemas/dailyRecord";
 import { withUserDateParse, withUserTimezone } from "@/lib/timezone";
-import { ValidationErrorResponse } from "@/schemas/common";
+import { ErrorResponse, ValidationErrorResponse } from "@/schemas/common";
+
+// GET: /records ユーザー_記録取得（特定日）
+export const GET = async (
+  request: NextRequest,
+  { params }: { params: { date: string } },
+) => {
+  const date = params.date;
+  if (!date) {
+    return NextResponse.json(
+      { status: "日付を指定してください" },
+      { status: 400 },
+    );
+  }
+
+  try {
+    const user = await getLoggedInUser(request);
+    const { recordedDate } = withUserDateParse(
+      { recordedDate: date },
+      ["recordedDate"],
+      user.timezone,
+    );
+    const dailyRecord = await prisma.dailyRecord.findUnique({
+      where: {
+        userId_recordedDate: {
+          userId: user.id,
+          recordedDate,
+        },
+      },
+    });
+
+    // frontで初回表示ロジックへ
+    if (!dailyRecord) {
+      return NextResponse.json(
+        { status: "記録を取得できませんでした" },
+        { status: 404 },
+      );
+    }
+
+    const converted = withUserTimezone(
+      dailyRecord,
+      [
+        "createdAt",
+        "updatedAt",
+        "recordedDate",
+        "commitStartDate",
+        "commitEndDate",
+      ],
+      user.timezone,
+    );
+
+    const safeDailyRecord: DailyRecord = converted;
+
+    return NextResponse.json<DailyRecordResponse>(
+      { dailyRecord: safeDailyRecord },
+      { status: 200 },
+    );
+  } catch (error) {
+    if (error instanceof Error) {
+      return NextResponse.json<ErrorResponse>(
+        { error: error.message },
+        { status: 400 },
+      );
+    }
+  }
+};
 
 // PUT: /records:date ユーザー_記録新規作成,取得,更新
-
-const toISODate = (d: Date) => d.toISOString().slice(0, 10); // "YYYY-MM-DD"
-
 export const PUT = async (
   request: NextRequest,
   { params }: { params: { date: string } },
